Refactor Orbital Tests
Given that Orbit and OrbitState are the only classes that really needed assistance in standing up their functionality with a bunch of helper functions, if those helper functions overlap significantly, maybe should just make a dedicated definition of them available, instead of hidden away.
